refactor(about): remove generic JSON Resume service
split into positions and education items services
